#  langs/i18n: Upgrade to go-i18n v2