Fixes #5242

Replaces #5243