@bep 2019-06-02
Relevant benchmarks (compared to master with go-i18n v1):

```bash
benchmark                                                    old ns/op     new ns/op     delta
BenchmarkI18nTranslate/all-present-4                         771           1918          +148.77%
BenchmarkI18nTranslate/present-in-default-4                  1373          52076         +3692.86%
BenchmarkI18nTranslate/present-in-current-4                  778           1908          +145.24%
BenchmarkI18nTranslate/missing-4                             1097          48754         +4344.30%
BenchmarkI18nTranslate/file-missing-4                        3658          3341          -8.67%
BenchmarkI18nTranslate/context-provided-4                    2054          2425          +18.06%
BenchmarkI18nTranslate/same-id-and-translation-4             798           1900          +138.10%
BenchmarkI18nTranslate/same-id-and-translation-default-4     1400          51376         +3569.71%
BenchmarkI18nTranslate/unknown-language-code-4               1817          3555          +95.65%

benchmark                                                    old allocs     new allocs     delta
BenchmarkI18nTranslate/all-present-4                         6              11             +83.33%
BenchmarkI18nTranslate/present-in-default-4                  10             208            +1980.00%
BenchmarkI18nTranslate/present-in-current-4                  6              11             +83.33%
BenchmarkI18nTranslate/missing-4                             8              199            +2387.50%
BenchmarkI18nTranslate/file-missing-4                        21             18             -14.29%
BenchmarkI18nTranslate/context-provided-4                    15             12             -20.00%
BenchmarkI18nTranslate/same-id-and-translation-4             6              11             +83.33%
BenchmarkI18nTranslate/same-id-and-translation-default-4     10             208            +1980.00%
BenchmarkI18nTranslate/unknown-language-code-4               13             21             +61.54%

benchmark                                                    old bytes     new bytes     delta
BenchmarkI18nTranslate/all-present-4                         152           528           +247.37%
BenchmarkI18nTranslate/present-in-default-4                  216           9667          +4375.46%
BenchmarkI18nTranslate/present-in-current-4                  152           528           +247.37%
BenchmarkI18nTranslate/missing-4                             152           9276          +6002.63%
BenchmarkI18nTranslate/file-missing-4                        600           608           +1.33%
BenchmarkI18nTranslate/context-provided-4                    704           552           -21.59%
BenchmarkI18nTranslate/same-id-and-translation-4             152           520           +242.11%
BenchmarkI18nTranslate/same-id-and-translation-default-4     216           9658          +4371.30%
BenchmarkI18nTranslate/unknown-language-code-4               696           1024          +47.13%

```