@bep 2019-06-02
Some benchmarks:

```bash
benchmark                                                    old ns/op     new ns/op     delta
BenchmarkI18nTranslate/all-present-4                         775           796           +2.71%
BenchmarkI18nTranslate/present-in-default-4                  1372          49923         +3538.70%
BenchmarkI18nTranslate/present-in-current-4                  777           797           +2.57%
BenchmarkI18nTranslate/missing-4                             1106          47469         +4191.95%
BenchmarkI18nTranslate/file-missing-4                        3729          2109          -43.44%
BenchmarkI18nTranslate/context-provided-4                    2087          1258          -39.72%
BenchmarkI18nTranslate/same-id-and-translation-4             806           792           -1.74%
BenchmarkI18nTranslate/same-id-and-translation-default-4     1413          49749         +3420.81%
BenchmarkI18nTranslate/unknown-language-code-4               1821          2073          +13.84%

benchmark                                                    old allocs     new allocs     delta
BenchmarkI18nTranslate/all-present-4                         6              4              -33.33%
BenchmarkI18nTranslate/present-in-default-4                  10             201            +1910.00%
BenchmarkI18nTranslate/present-in-current-4                  6              4              -33.33%
BenchmarkI18nTranslate/missing-4                             8              192            +2300.00%
BenchmarkI18nTranslate/file-missing-4                        21             11             -47.62%
BenchmarkI18nTranslate/context-provided-4                    15             5              -66.67%
BenchmarkI18nTranslate/same-id-and-translation-4             6              4              -33.33%
BenchmarkI18nTranslate/same-id-and-translation-default-4     10             201            +1910.00%
BenchmarkI18nTranslate/unknown-language-code-4               13             12             -7.69%

benchmark                                                    old bytes     new bytes     delta
BenchmarkI18nTranslate/all-present-4                         152           176           +15.79%
BenchmarkI18nTranslate/present-in-default-4                  216           9315          +4212.50%
BenchmarkI18nTranslate/present-in-current-4                  152           176           +15.79%
BenchmarkI18nTranslate/missing-4                             152           8924          +5771.05%
BenchmarkI18nTranslate/file-missing-4                        600           256           -57.33%
BenchmarkI18nTranslate/context-provided-4                    704           200           -71.59%
BenchmarkI18nTranslate/same-id-and-translation-4             152           165           +8.55%
BenchmarkI18nTranslate/same-id-and-translation-default-4     216           9304          +4207.41%
BenchmarkI18nTranslate/unknown-language-code-4               696           640           -8.05%
```

I don't see any obviously wrong with my implementation. One curious example:

```bash
BenchmarkI18nTranslate/present-in-default-4                  1372          49923         +3538.70%
```

Which is lookup for id `hello` for language `es`. That string does not exist in `es`, but does exist in default language `en`.

How does the above correspond to your benchmarks, @nicksnyder ?