@bep 2020-09-29
These are benchmarks more to my liking:

```bash
name                                              old time/op    new time/op    delta
I18nTranslate/all-present-16                         768ns ± 0%     520ns ± 1%   -32.40%  (p=0.029 n=4+4)
I18nTranslate/present-in-default-16                 1.40µs ± 0%    1.23µs ± 0%   -11.80%  (p=0.029 n=4+4)
I18nTranslate/present-in-current-16                  768ns ± 1%     512ns ± 2%   -33.36%  (p=0.029 n=4+4)
I18nTranslate/missing-16                            1.08µs ± 0%    1.18µs ± 0%    +9.04%  (p=0.029 n=4+4)
I18nTranslate/file-missing-16                       4.24µs ± 0%    2.51µs ± 0%   -40.88%  (p=0.029 n=4+4)
I18nTranslate/context-provided-16                   2.04µs ± 0%    1.32µs ± 0%   -35.08%  (p=0.029 n=4+4)
I18nTranslate/same-id-and-translation-16             800ns ± 0%     506ns ± 1%   -36.78%  (p=0.029 n=4+4)
I18nTranslate/same-id-and-translation-default-16    1.43µs ± 0%    1.23µs ± 0%   -13.73%  (p=0.029 n=4+4)
I18nTranslate/unknown-language-code-16              1.70µs ± 0%    2.22µs ± 0%   +30.59%  (p=0.029 n=4+4)

name                                              old alloc/op   new alloc/op   delta
I18nTranslate/all-present-16                          152B ± 0%        0B       -100.00%  (p=0.029 n=4+4)
I18nTranslate/present-in-default-16                   216B ± 0%      112B ± 0%   -48.15%  (p=0.029 n=4+4)
I18nTranslate/present-in-current-16                   152B ± 0%        0B       -100.00%  (p=0.029 n=4+4)
I18nTranslate/missing-16                              152B ± 0%      112B ± 0%   -26.32%  (p=0.029 n=4+4)
I18nTranslate/file-missing-16                         600B ± 0%      288B ± 0%   -52.00%  (p=0.029 n=4+4)
I18nTranslate/context-provided-16                     704B ± 0%      200B ± 0%   -71.59%  (p=0.029 n=4+4)
I18nTranslate/same-id-and-translation-16              152B ± 0%        0B       -100.00%  (p=0.029 n=4+4)
I18nTranslate/same-id-and-translation-default-16      216B ± 0%      112B ± 0%   -48.15%  (p=0.029 n=4+4)
I18nTranslate/unknown-language-code-16                696B ± 0%      640B ± 0%    -8.05%  (p=0.029 n=4+4)

name                                              old allocs/op  new allocs/op  delta
I18nTranslate/all-present-16                          6.00 ± 0%      0.00       -100.00%  (p=0.029 n=4+4)
I18nTranslate/present-in-default-16                   10.0 ± 0%       5.0 ± 0%   -50.00%  (p=0.029 n=4+4)
I18nTranslate/present-in-current-16                   6.00 ± 0%      0.00       -100.00%  (p=0.029 n=4+4)
I18nTranslate/missing-16                              8.00 ± 0%      5.00 ± 0%   -37.50%  (p=0.029 n=4+4)
I18nTranslate/file-missing-16                         21.0 ± 0%      11.0 ± 0%   -47.62%  (p=0.029 n=4+4)
I18nTranslate/context-provided-16                     15.0 ± 0%       5.0 ± 0%   -66.67%  (p=0.029 n=4+4)
I18nTranslate/same-id-and-translation-16              6.00 ± 0%      0.00       -100.00%  (p=0.029 n=4+4)
I18nTranslate/same-id-and-translation-default-16      10.0 ± 0%       5.0 ± 0%   -50.00%  (p=0.029 n=4+4)
I18nTranslate/unknown-language-code-16                13.0 ± 0%      12.0 ± 0%    -7.69%  (p=0.029 n=4+4)
```